add separate bookkeeping table component
